requirements.txt
termcolor==2.4.0
requests
pyyaml
